refactor(js/plugins/vertexai): extract rerankers
